requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2.20
django-filter==25.1
djangorestframework==3.16.0
exceptiongroup==1.2.2
importlib_metadata==8.6.1
iniconfig==2.1.0
isort==6.0.1
Markdown==3.7
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.11.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.0
tzdata==2025.2
zipp==3.21.0
